core: pathfinding: split PathfindingResult into data and conversion
