+++ core/src/main/java/fr/sncf/osrd/api/pathfinding/response/PathfindingResult.java
@@ -37,239 +37,6 @@ public class PathfindingResult {
     public LineString geographic = null;
     public LineString schematic = null;
 
-    public List<Warning> warnings;
-
-    /**
-     * pathfindingResultMake is used to create a class PathfindingResult with a good format to be sent to the middleware
-     * during this function, we add geometric information in the PathfindingResult
-     *
-     * @param path contains the pathfinding's result
-     */
-    public static PathfindingResult make(
-            Pathfinding.Result<SignalingRoute> path,
-            SignalingInfra infra,
-            WarningRecorderImpl warningRecorder
-    ) {
-        var res = new PathfindingResult();
-
-        // Builds a mapping between routes and all user defined waypoints on the route
-        var userDefinedWaypointsPerRoute = HashMultimap.<SignalingRoute, Double>create();
-        for (var waypoint : path.waypoints())
-            userDefinedWaypointsPerRoute.put(waypoint.edge(), waypoint.offset());
-
-        for (var signalingRouteEdgeRange : path.ranges()) {
-            if (signalingRouteEdgeRange.start() < signalingRouteEdgeRange.end())
-                res.routePaths.add(makeRouteResult(signalingRouteEdgeRange));
-            var waypoints = getWaypointsOnRoute(
-                    signalingRouteEdgeRange,
-                    userDefinedWaypointsPerRoute.get(signalingRouteEdgeRange.edge())
-            );
-            for (var waypoint : waypoints)
-                res.addStep(waypoint);
-        }
-        res.addGeometry(infra);
-        res.warnings = warningRecorder.warnings;
-        return res;
-    }
-
-    /** Adds all waypoints on the route range, both suggestions (OP) and user defined waypoints */
-    public static List<PathWaypointResult> getWaypointsOnRoute(
-            Pathfinding.EdgeRange<SignalingRoute> routeRange,
-            Set<Double> userDefinedWaypointOffsets
-    ) {
-        var res = new ArrayList<PathWaypointResult>();
-
-        // We need a mutable copy to remove elements as we find them
-        var waypointsOffsetsList = new ArrayList<>(userDefinedWaypointOffsets);
-
-        double offset = 0;
-        for (var range : routeRange.edge().getInfraRoute().getTrackRanges()) {
-            if (!(range.track.getEdge() instanceof TrackSection trackSection))
-                continue;
-
-            if (offset > routeRange.end())
-                break;
-
-            if (offset + range.getLength() >= routeRange.start()) {
-                // Truncate the range to match the part of the route we use
-                var truncatedRange = truncateTrackRange(range, offset, routeRange);
-                if (truncatedRange != null) {
-                    // List all waypoints on the track range in a tree map, with offsets from the range start as key
-                    var waypoints = new ArrayList<PathWaypointResult>();
-
-                    // Add operational points as optional waypoints
-                    for (var op : truncatedRange.getOperationalPoints())
-                        waypoints.add(PathWaypointResult.suggestion(op.element(), trackSection, op.offset()));
-
-                    // Add user defined waypoints
-                    var truncatedRangeOffset = offset + Math.abs(truncatedRange.getStart() - range.getStart());
-                    for (int i = 0; i < waypointsOffsetsList.size(); i++) {
-                        var trackRangeOffset = waypointsOffsetsList.get(i) - truncatedRangeOffset;
-
-                        // We can have tiny differences here because we accumulate offsets in a different way
-                        if (Math.abs(trackRangeOffset - truncatedRange.getLength()) < POSITION_EPSILON)
-                            trackRangeOffset = truncatedRange.getLength();
-                        if (Math.abs(trackRangeOffset) < POSITION_EPSILON)
-                            trackRangeOffset = 0;
-
-                        if (trackRangeOffset >= 0 && trackRangeOffset <= truncatedRange.getLength()) {
-                            var location = truncatedRange.offsetLocation(trackRangeOffset);
-                            waypoints.add(PathWaypointResult.userDefined(location, trackRangeOffset));
-                            waypointsOffsetsList.remove(i--); // Avoids duplicates on track transitions
-                        }
-                    }
-
-                    // Adds all waypoints in order
-                    waypoints.sort(Comparator.comparingDouble(x -> x.trackRangeOffset));
-                    res.addAll(waypoints);
-                }
-            }
-
-            offset += range.getLength();
-        }
-        assert waypointsOffsetsList.isEmpty();
-        return res;
-    }
-
-    /** Truncates a track range view so that it's limited to the route range */
-    private static TrackRangeView truncateTrackRange(
-            TrackRangeView range,
-            double offset,
-            Pathfinding.EdgeRange<SignalingRoute> routeRange
-    ) {
-        var truncatedRange = range;
-        if (offset < routeRange.start()) {
-            var truncateLength = routeRange.start() - offset;
-            if (truncateLength > truncatedRange.getLength() + POSITION_EPSILON)
-                return null;
-            truncatedRange = truncatedRange.truncateBeginByLength(truncateLength);
-        }
-        if (offset + range.getLength() > routeRange.end()) {
-            var truncateLength = offset + range.getLength() - routeRange.end();
-            if (truncateLength > truncatedRange.getLength() + POSITION_EPSILON)
-                return null;
-            truncatedRange = truncatedRange.truncateEndByLength(truncateLength);
-        }
-        return truncatedRange;
-    }
-
-    /** Adds a single route to the result, including waypoints present on the route */
-    private static RoutePathResult makeRouteResult(
-            Pathfinding.EdgeRange<SignalingRoute> element
-    ) {
-        var routeResult = new RoutePathResult(
-                new RJSObjectRef<>(element.edge().getInfraRoute().getID(), "Route"),
-                element.edge().getSignalingType()
-        );
-        double offset = 0;
-        for (var range : element.edge().getInfraRoute().getTrackRanges()) {
-            if (!(range.track.getEdge() instanceof TrackSection trackSection))
-                continue;
-
-            // Truncate the ranges to match the part of the route we use
-            var truncatedRange = truncateTrackRange(range, offset, element);
-            offset += range.getLength();
-            if (truncatedRange == null)
-                continue;
-
-            // Add the track ranges to the result
-            routeResult.trackSections.add(new DirTrackRange(
-                    trackSection.getID(),
-                    truncatedRange.getStart(),
-                    truncatedRange.getStop()
-            ));
-        }
-        return routeResult;
-    }
-
-    /** Generates the path geometry */
-    void addGeometry(SignalingInfra infra) {
-        var geoList = new ArrayList<LineString>();
-        var schList = new ArrayList<LineString>();
-
-        DirTrackRange previousTrack = null;
-        double previousBegin = 0;
-        double previousEnd = 0;
-
-        for (var routePath : routePaths) {
-            for (var trackSection : routePath.trackSections) {
-
-                if (trackSection.getBegin() == trackSection.getEnd())
-                    continue;
-
-                if (previousTrack == null) {
-                    previousTrack = trackSection;
-                    previousBegin = trackSection.getBegin();
-                    previousEnd = trackSection.getEnd();
-                    continue;
-                }
-
-                if (previousTrack.trackSection.id.id.compareTo(trackSection.trackSection.id.id) != 0) {
-                    if (Double.compare(previousBegin, previousEnd) != 0) {
-                        var track = RJSObjectParsing.getTrackSection(previousTrack.trackSection, infra);
-                        sliceAndAdd(geoList, track.getGeo(), previousBegin, previousEnd, track.getLength());
-                        sliceAndAdd(schList, track.getSch(), previousBegin, previousEnd, track.getLength());
-                    }
-                    previousTrack = trackSection;
-                    previousBegin = trackSection.getBegin();
-                }
-                previousEnd = trackSection.getEnd();
-            }
-        }
-
-        assert previousTrack != null;
-        var track = RJSObjectParsing.getTrackSection(previousTrack.trackSection, infra);
-        sliceAndAdd(geoList, track.getGeo(), previousBegin, previousEnd, track.getLength());
-        sliceAndAdd(schList, track.getSch(), previousBegin, previousEnd, track.getLength());
-
-        geographic = concatenate(geoList);
-        schematic = concatenate(schList);
-    }
-
-    /** Concatenates a list of LineString into a single LineString.
-     * If not enough values are present, we return the default [0, 1] line. */
-    private LineString concatenate(List<LineString> list) {
-        if (list.size() >= 2)
-            return LineString.concatenate(list);
-        else if (list.size() == 1)
-            return list.get(0);
-        else
-            return LineString.make(
-                    new double[] {0., 1.},
-                    new double[] {0., 1.}
-            );
-    }
-
-    /** If the lineString isn't null, slice it from previousBegin to previousEnd and add it to res */
-    private static void sliceAndAdd(
-            List<LineString> res,
-            LineString lineString,
-            double previousBegin,
-            double previousEnd,
-            double trackLength
-    ) {
-        if (lineString == null)
-            return;
-        if (trackLength == 0) {
-            assert previousBegin == 0;
-            assert previousEnd == 0;
-            res.add(lineString);
-        } else
-            res.add(lineString.slice(previousBegin / trackLength, previousEnd / trackLength));
-    }
-
-    /** Adds a single waypoint to the result */
-    void addStep(PathWaypointResult newStep) {
-        if (pathWaypoints.isEmpty()) {
-            pathWaypoints.add(newStep);
-            return;
-        }
-        var lastStep = pathWaypoints.get(pathWaypoints.size() - 1);
-        if (lastStep.isDuplicate(newStep)) {
-            lastStep.merge(newStep);
-            return;
-        }
-        pathWaypoints.add(newStep);
-    }
+    public List<Warning> warnings = null;
 }
